Consider all PR check runs and commit statuses before merging
Previous to this commit, if a single check suite passed on the PR, the
PR would be merged. This is incorrect since a check suite only
represents a grouping of check runs by a single GitHub app.

This commit changes Magic Mirror to consider all check runs (created by
GitHub apps) and commit statuses (non-GitHub apps) which are required by
the branch protection configuration. If there are no required check runs
or commit statuses on the branch, Magic Mirror merges the PR right away.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>